backend/requirements.txt
fastapi
uvicorn
motor
redis
python-multipart
passlib[bcrypt]
bcrypt
python-jose[cryptography]
pydantic
